converting vector and matrix to half instead of float for better gpu performance

+++ SRC/AABB.hpp
@@ -37,10 +37,10 @@ public:
 	}
 	const bool ClipLine(const AABB& aabb, const Vector3& start, const Vector3& end, float& f_low, float& f_high) const
 	{
-		float3 f_dim_low = (aabb.min.xyz - start.xyz) / (end.xyz - start.xyz);
-		float3 f_dim_high = (aabb.max.xyz - start.xyz) / (end.xyz - start.xyz);
+		half3 f_dim_low = (aabb.min.xyz - start.xyz) / (end.xyz - start.xyz);
+		half3 f_dim_high = (aabb.max.xyz - start.xyz) / (end.xyz - start.xyz);
 		
-		int3 comp = f_dim_high < f_dim_low;
+		short3 comp = f_dim_high < f_dim_low;
 		for(int i = 0; i < 3; i++)
 			if (comp[i])
 			{
@@ -48,11 +48,11 @@ public:
 				f_dim_high[i] = f_dim_low[i];
 				f_dim_low[i] = temp;
 			}
-		int3 first1 = f_dim_high < f_low;
-		int3 second1 = f_dim_low > f_high;
+		short3 first1 = f_dim_high < f_low;
+		short3 second1 = f_dim_low > f_high;
 		
-		int3 first2 = f_dim_low > f_low;
-		int3 second2 = f_dim_high < f_high;
+		short3 first2 = f_dim_low > f_low;
+		short3 second2 = f_dim_high < f_high;
 		for(int i = 0; i < 3; i++)
 		{
 			if (first1[i] || second1[i])
@@ -70,8 +70,8 @@ public:
 
 	const bool AABBIntersection(const AABB &other) const
 	{
-		int3 first = min.xyz > other.max.xyz;
-		int3 second = max.xyz < other.min.xyz;
+		short3 first = min.xyz > other.max.xyz;
+		short3 second = max.xyz < other.min.xyz;
 		for(char i = 0; i < 3; i++)
 		{
 			if(first[i])
